specific counters for each DDR
