+++ src/TraceParser.cc
@@ -36,8 +36,6 @@
 //
 #include "TraceParser.h"
 
-#include "Counters.h"
-
 using namespace DRAMPower;
 using namespace std;
