+++ src/Utils.h
@@ -57,4 +57,6 @@ T fromString(const std::string& s,
   return t;
 }
 
+
+
 #endif /* UTILS_H */
